% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos_cnae.R
\encoding{UTF-8}
\name{codigos_cnae}
\alias{codigos_cnae}
\title{Códigos Classificação Nacional de Atividades Econômicas}
\usage{
codigos_cnae(tabela, campo, nivel, add_nomes = F)
}
\arguments{
\item{tabela}{A tabela para adicionar outros níveis CNAE.}

\item{campo}{A coluna com os códigos subclasse da CNAE.}

\item{nivel}{O nível CNAE desejado. Deve ser "seção", "divisão", "grupo" ou "classe".}

\item{add_nomes}{Logical: deseja adicionar uma coluna com os nomes?}
}
\value{
Uma nova coluna com os códigos CNAE em nível mais agrupados a partir de subclasse.
}
\description{
Códigos Classificação Nacional de Atividades Econômicas
}
\examples{
df1 <- data.frame(col1 = c(9609201, 111302))
df2 <- codigos_cnae(df1, col1, "seção")
df2

df3 <- data.frame(col1 = c("9609201", "0111302"))
df4 <- codigos_cnae(df3, col1, "classe")
df4
}
