% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos_cnae_niveis.R
\encoding{UTF-8}
\name{codigos_cnae_niveis}
\alias{codigos_cnae_niveis}
\title{Códigos Classificação Nacional de Atividades Econômicas}
\usage{
codigos_cnae_niveis(tabela, campo, nivel, add_nomes = F)
}
\arguments{
\item{tabela}{Dataframe: a tabela para adicionar outros níveis CNAE.}

\item{campo}{Caracter: a coluna com os códigos subclasse da CNAE.}

\item{nivel}{Caracter: o nível CNAE desejado. Deve ser "seção", "divisão", "grupo" ou "classe".}

\item{add_nomes}{Logical: deseja adicionar uma coluna com os nomes?}
}
\value{
Uma nova coluna com os códigos CNAE em nível mais agrupados a partir de subclasse.
}
\description{
Códigos Classificação Nacional de Atividades Econômicas
}
\examples{
df1 <- data.frame(cnae_subclasse = c(9609201, 111302))
df2 <- codigos_cnae_niveis(df1, cnae_subclasse, "seção")
df2

df3 <- data.frame(cnae_subclasse = c("9609201", "0111302"))
df4 <- codigos_cnae_niveis(df3, cnae_subclasse, "classe", add_nomes = TRUE)
df4
}
