% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos_cnae_nivel.R
\encoding{UTF-8}
\name{codigos_cnae_nivel}
\alias{codigos_cnae_nivel}
\title{Códigos Classificação Nacional de Atividades Econômicas - Níveis}
\usage{
codigos_cnae_nivel(tabela, campo, nivel, add_nomes = FALSE)
}
\arguments{
\item{tabela}{Dataframe: a tabela para adicionar outros níveis CNAE.}

\item{campo}{Caractere: a coluna com os códigos subclasse da CNAE.}

\item{nivel}{Caractere: o nível CNAE desejado. Deve ser "seção", "divisão", "grupo" ou "classe".}

\item{add_nomes}{Logical: deseja adicionar uma coluna com os nomes?}
}
\value{
Uma nova coluna com os códigos CNAE em nível mais agrupados a partir de subclasse.
}
\description{
Procura em dataframe por coluna com os códigos da CNAE subclasse e adiciona coluna extra contendo
os códigos CNAE equivalentes em nível superior (classe, grupo, divisão ou seção), conforme escolha.
}
\examples{
df1 <- data.frame(cnae_subclasse = c(9609201, 111302))
df2 <- codigos_cnae_nivel(df1, "cnae_subclasse", "classe")
df2

df3 <- data.frame(cnae_subclasse = c("9609201", "0111302"))
df4 <- codigos_cnae_nivel(df3, "cnae_subclasse", "seção", add_nomes = TRUE)
df4
}
