% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos_municipio_uf.R
\encoding{UTF-8}
\name{codigos_municipio_uf}
\alias{codigos_municipio_uf}
\title{Códigos Municípios - Unidades Federativas Brasileiras}
\usage{
codigos_municipio_uf(tabela, campo, add_nomes = FALSE)
}
\arguments{
\item{tabela}{Dataframe: a tabela para adicionar os códigos das unidades federativas}

\item{campo}{Caractere: a coluna com os códigos dos municípios brasileiros}

\item{add_nomes}{Logical: deseja adicionar uma coluna com os nomes das unidades federativas?}
}
\value{
Adiciona uma coluna com os códigos das unidades federativas brasileiras em uma tabela que possui apenas os códigos dos municípios.
}
\description{
Procura em dataframe por coluna com códigos dos municípios brasileiros estabelecidos pelo IBGE
e adiciona uma coluna com os códigos das unidades federativas relacionadas.
}
\examples{
df1 <- data.frame(codigos_municipios = c("5219704", "3123858", "2111532"))
df2 <- codigos_municipio_uf(df1, "codigos_municipios")
df2

df3 <- data.frame(codigos_municipios = c(521970, 312385, 211153))
df4 <- codigos_municipio_uf(df3, "codigos_municipios", add_nomes = TRUE)
df4
}
