% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos_ncm_cnae.R
\name{codigos_ncm_cnae}
\alias{codigos_ncm_cnae}
\title{Códigos Nomeclatura Comum do Mercosul - Classificação Brasileira de Atividades Econômicas (classe)}
\usage{
codigos_ncm_cnae(tabela, campo, add_nomes = FALSE)
}
\arguments{
\item{tabela}{Dataframe: a tabela para adicionar os códigos das classes CNAE.}

\item{campo}{Caractere: a coluna com os códigos NCM.}

\item{add_nomes}{Logical: deseja adicionar uma coluna com os nomes das classes CNAE?}
}
\value{
Uma nova coluna com os códigos das classes CNAE equivalentes aos códigos NCM.
}
\description{
Em um dataframe que possui uma coluna com os códigos NCM, adiciona uma coluna extra com os códigos
das classes CNAE equivalentes. Adiciona ainda os nomes destas, se desejar. A tradução é realizada de acordo com
a correspondência NCM 2012 x CNAE 2.0 divulgada pelo
\href{https://concla.ibge.gov.br/classificacoes/correspondencias/atividades-economicas.html}{IBGE}.
}
\examples{
df1 <- data.frame(codigos_ncm = c("01012100", "35052000", "84459030"))
df2 <- codigos_ncm_cnae(df1, "codigos_ncm")
df2

df3 <- data.frame(codigos_ncm = c(1012100, 35052000, 84459030))
df4 <- codigos_ncm_cnae(df3, "codigos_ncm", add_nomes = TRUE)
df4
}
