% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomeia_cnae.R
\encoding{UTF-8}
\name{nomeia_cnae}
\alias{nomeia_cnae}
\title{Nomes Classificação Nacional de Atividades Econômicas - Subclasse}
\usage{
nomeia_cnae(tabela, campo, nivel = "subclasse")
}
\arguments{
\item{tabela}{Dataframe: a tabela para adicionar os nomes das subclasses da CNAE}

\item{campo}{Caractere: a coluna com os códigos das subclasses da CNAE}

\item{nivel}{Caractere: o nível CNAE. Deve ser "subclasse", "classe", "grupo", "divisão" ou "seção".}
}
\value{
Adiciona uma coluna com as descrições das subclasses da CNAE em uma tabela que possui apenas os códigos.
}
\description{
Procura em dataframe por coluna com códigos CNAE subclasse e adiciona uma coluna com os nomes relacionados.
}
\examples{
df1 <- data.frame(codigos_cnae_subclasse = c("0111301", "4520008", "9900800"))
df2 <- nomeia_cnae(df1, "codigos_cnae_subclasse", "subclasse")
df2

df3 <- data.frame(codigos_cnae_grupo = c(11, 452, 990))
df4 <- nomeia_cnae(df3, "codigos_cnae_grupo", "grupo")
df4
}
