% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomes_cbo.R
\encoding{UTF-8}
\name{nomes_cbo}
\alias{nomes_cbo}
\title{Nomes Classificação Brasileira de Ocupação}
\usage{
nomes_cbo(tabela, campo)
}
\arguments{
\item{tabela}{Dataframe: a tabela para adicionar os nomes das CBOs}

\item{campo}{Caractere: a coluna com os códigos CBO}
}
\value{
Adiciona uma coluna com as descrições das CBOs em uma tabela que possui apenas os códigos.
}
\description{
Procura em dataframe por coluna com códigos CBO e adiciona uma coluna com os nomes relacionados.
}
\examples{
df1 <- data.frame(codigos_cbo = c("010105", "515220", "992115"))
df2 <- nomes_cbo(df1, "codigos_cbo")
df2

df3 <- data.frame(codigos_cbo = c(010105, 515220, 992115))
df4 <- nomes_cbo(df3, "codigos_cbo")
df4
}
