% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomes_cnae_classe.R
\encoding{UTF-8}
\name{nomes_cnae_classe}
\alias{nomes_cnae_classe}
\title{Nomes Classificação Nacional de Atividades Econômicas - Classe}
\usage{
nomes_cnae_classe(tabela, campo)
}
\arguments{
\item{tabela}{Dataframe: a tabela para adicionar os nomes das classes da CNAE}

\item{campo}{Caractere: a coluna com os códigos das classes da CNAE}
}
\value{
Adiciona uma coluna com as descrições das classes da CNAE em uma tabela que possui apenas os códigos.
}
\description{
Procura em dataframe por coluna com códigos CNAE classe e adiciona uma coluna com os nomes relacionados.
}
\examples{
df1 <- data.frame(codigos_cnae_classes = c("01113", "45200", "99008"))
df2 <- nomes_cnae_classe(df1, "codigos_cnae_classes")
df2

df3 <- data.frame(codigos_cnae_classes = c(1113, 45200, 99008))
df4 <- nomes_cnae_classe(df3, "codigos_cnae_classes")
df4
}
