% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomes_cnae_grupo.R
\encoding{UTF-8}
\name{nomes_cnae_grupo}
\alias{nomes_cnae_grupo}
\title{Nomes Classificação Nacional de Atividades Econômicas - Grupo}
\usage{
nomes_cnae_grupo(tabela, campo)
}
\arguments{
\item{tabela}{Dataframe: a tabela para adicionar os nomes das grupos da CNAE}

\item{campo}{Caractere: a coluna com os códigos das grupos da CNAE}
}
\value{
Adiciona uma coluna com as descrições das grupos da CNAE em uma tabela que possui apenas os códigos.
}
\description{
Procura em dataframe por coluna com códigos CNAE grupo e adiciona uma coluna com os nomes relacionados.
}
\examples{
df1 <- data.frame(codigos_cnae_grupo = c("011", "452", "990"))
df2 <- nomes_cnae_grupo(df1, "codigos_cnae_grupo")
df2

df3 <- data.frame(codigos_cnae_grupo = c(11, 452, 990))
df4 <- nomes_cnae_grupo(df3, "codigos_cnae_grupo")
df4
}
