% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomes_cnae_secao.R
\encoding{UTF-8}
\name{nomes_cnae_secao}
\alias{nomes_cnae_secao}
\title{Nomes Classificação Nacional de Atividades Econômicas - Seção}
\usage{
nomes_cnae_secao(tabela, campo)
}
\arguments{
\item{tabela}{Dataframe: a tabela para adicionar os nomes das seções da CNAE}

\item{campo}{Caractere: a coluna com os códigos das seções da CNAE}
}
\value{
Adiciona uma coluna com as descrições das seções da CNAE em uma tabela que possui apenas os códigos.
}
\description{
Procura em dataframe por coluna com códigos CNAE seção e adiciona uma coluna com os nomes relacionados.
}
\examples{
df1 <- data.frame(codigos_cnae_secao = c("A", "G", "U"))
df2 <- nomes_cnae_secao(df1, "codigos_cnae_secao")
df2
}
