% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomes_cnae_subclasse.R
\name{nomes_cnae_subclasse}
\alias{nomes_cnae_subclasse}
\title{Nomes Classificação Nacional de Atividades Econômicas - Subclasse}
\usage{
nomes_cnae_subclasse(tabela, campo)
}
\arguments{
\item{tabela}{Data.frame: a tabela para adicionar os nomes das subclasses da CNAE}

\item{campo}{Caracter: a coluna com os códigos das subclasses da CNAE}
}
\value{
Adiciona uma coluna com as descrições das subclasses da CNAE em uma tabela que possui apenas os códigos.
}
\description{
Nomes Classificação Nacional de Atividades Econômicas - Subclasse
}
\examples{
df1 <- data.frame(col1 = c("0111301","4520008","9900800"))
df2 <- nomes_cnae_subclasse(df1, "col1")
df2

df3 <- data.frame(col1 = c(111301,4520008,9900800))
df4 <- nomes_cnae_subclasse(df1, "col1")
df4
}
