% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomes_uf.R
\encoding{UTF-8}
\name{nomes_uf}
\alias{nomes_uf}
\title{Nomes Unidades Federativas Brasileiras}
\usage{
nomes_uf(tabela, campo, nome = TRUE, sigla = FALSE)
}
\arguments{
\item{tabela}{Dataframe: a tabela para adicionar os nomes os siglas das unidades federativas brasileiras}

\item{campo}{Caractere: a coluna com os códigos das unidades federativas brasileiras}

\item{nome}{Logical: deseja que seja adiconado os nomes das UFs?}

\item{sigla}{Logical: deseja que seja adiconado as siglas das UFs?}
}
\value{
Adiciona uma coluna com os nomes ou as siglas das unidades federativas brasileiras em uma tabela que possui apenas os códigos.
}
\description{
Procura em dataframe por coluna com códigos das unidades federativas brasileiras estabelecidos pelo IBGE
e adiciona uma coluna com os nomes ou siglas relacionadas.
}
\examples{
df1 <- data.frame(codigos_uf = c("11", "22", "43"))
df2 <- nomes_uf(df1, "codigos_uf")
df2

df3 <- data.frame(codigos_uf = c(11, 22, 43))
df4 <- nomes_uf(df3, "codigos_uf", nome = FALSE, sigla = TRUE)
df4
}
