% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traduz_cnae_nivel.R
\encoding{UTF-8}
\name{traduz_cnae_nivel}
\alias{traduz_cnae_nivel}
\title{Códigos Classificação Nacional de Atividades Econômicas - Níveis}
\usage{
traduz_cnae_nivel(
  tabela,
  campo,
  nivel_input = "subclasse",
  nivel_output = "seção",
  add_nomes = FALSE
)
}
\arguments{
\item{tabela}{Dataframe: a tabela para adicionar outros níveis CNAE.}

\item{campo}{Caractere: a coluna com os códigos da CNAE atuais para input.}

\item{nivel_input}{Caractere: o nível CNAE atual. Deve ser "subclasse", "classe", "grupo" ou "divisão".}

\item{nivel_output}{Caractere: o nível CNAE desejado. Deve ser "classe", "grupo", "divisão" ou "seção".}

\item{add_nomes}{Logical: deseja adicionar uma coluna com os nomes?}
}
\value{
Uma nova coluna com os códigos CNAE em nível mais agrupados. Se desejar, uma outra coluna com os nomes.
}
\description{
Procura em dataframe por coluna com os códigos da CNAE subclasse e adiciona coluna extra contendo
os códigos CNAE equivalentes em nível superior (classe, grupo, divisão ou seção), conforme escolha.
}
\examples{
df1 <- data.frame(cnae_subclasse = c(111302, 9609201))
df2 <- traduz_cnae_nivel(df1, "cnae_subclasse")
df2

df3 <- data.frame(cnae_grupo = c("011", "960"))
df4 <- traduz_cnae_nivel(df3, "cnae_grupo", "grupo", "divisão", TRUE)
df4
}
