% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos_cnae.R
\encoding{UTF-8}
\name{codigos_CNAE}
\alias{codigos_CNAE}
\title{Códigos Classificação Nacional de Atividades Econômicas}
\usage{
codigos_CNAE(tabela, campo, nivel, add_nomes = F)
}
\arguments{
\item{tabela}{Data.frame: a tabela para adicionar outros níveis CNAE}

\item{campo}{Caracter: a coluna com os códigos subclasse}

\item{nivel}{Caracter: o nível CNAE desejado. Deve ser "Seção", "Divisão", "Grupo" ou "Classe"}

\item{add_nomes}{Logical: deseja adicionar uma coluna com os nomes?}
}
\value{
Uma nova coluna com os códigos CNAE em algum nível mais agrupados.
}
\description{
Códigos Classificação Nacional de Atividades Econômicas
}
