% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomes_cbo.R
\encoding{UTF-8}
\name{nomes_cbo}
\alias{nomes_cbo}
\title{Nomes Classificação Brasileira de Ocupação}
\usage{
nomes_cbo(tabela, campo)
}
\arguments{
\item{tabela}{Data.frame: a tabela para adicionar os nomes das CBOs}

\item{campo}{Caracter: a coluna com os códigos CBO}
}
\value{
Adiciona uma coluna com as descrições das CBOs em uma tabela que possui apenas os códigos.
}
\description{
Nomes Classificação Brasileira de Ocupação
}
\examples{
df1 <- data.frame(col1 = c("010105", "515220", "992115"))
df2 <- nomes_cbo(df1, "col1")
df2

df3 <- data.frame(col1 = c(010105, 515220, 992115))
df4 <- nomes_cbo(df3, "col1")
df4
}
