% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomes_cnae_divisao.R
\encoding{UTF-8}
\name{nomes_cnae_divisao}
\alias{nomes_cnae_divisao}
\title{Nomes Classificação Nacional de Atividades Econômicas - Divisão}
\usage{
nomes_cnae_divisao(tabela, campo)
}
\arguments{
\item{tabela}{Data.frame: a tabela para adicionar os nomes das divisões da CNAE}

\item{campo}{Caracter: a coluna com os códigos das divisões da CNAE}
}
\value{
Adiciona uma coluna com as descrições das divisões da CNAE em uma tabela que possui apenas os códigos.
}
\description{
Nomes Classificação Nacional de Atividades Econômicas - Divisão
}
\examples{
df1 <- data.frame(codigos_cnae_divisao = c("01", "45", "99"))
df2 <- nomes_cnae_divisao(df1, "codigos_cnae_divisao")
df2

df3 <- data.frame(codigos_cnae_divisao = c(1, 45, 99))
df4 <- nomes_cnae_divisao(df3, "codigos_cnae_divisao")
df4
}
