% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomes_cnae_grupo.R
\encoding{UTF-8}
\name{nomes_cnae_grupo}
\alias{nomes_cnae_grupo}
\title{Nomes Classificação Nacional de Atividades Econômicas - Grupo}
\usage{
nomes_cnae_grupo(tabela, campo)
}
\arguments{
\item{tabela}{Data.frame: a tabela para adicionar os nomes das grupos da CNAE}

\item{campo}{Caracter: a coluna com os códigos das grupos da CNAE}
}
\value{
Adiciona uma coluna com as descrições das grupos da CNAE em uma tabela que possui apenas os códigos.
}
\description{
Nomes Classificação Nacional de Atividades Econômicas - Grupo
}
\examples{
df1 <- data.frame(col1 = c("011","452","990"))
df2 <- nomes_cnae_grupo(df1, "col1")
df2

df3 <- data.frame(col1 = c(11,452,990))
df4 <- nomes_cnae_grupo(df3, "col1")
df4
}
