% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomes_cnae_secao.R
\encoding{UTF-8}
\name{nomes_cnae_secao}
\alias{nomes_cnae_secao}
\title{Nomes Classificação Nacional de Atividades Econômicas - Seção}
\usage{
nomes_cnae_secao(tabela, campo)
}
\arguments{
\item{tabela}{A tabela para adicionar os nomes das seções da CNAE}

\item{campo}{A coluna com os códigos das seções da CNAE}
}
\value{
Adiciona uma coluna com as descrições das seções da CNAE em uma tabela que possui apenas os códigos.
}
\description{
Nomes Classificação Nacional de Atividades Econômicas - Seção
}
\examples{
df1 <- data.frame(col1 = c("A", "G", "U"))
df2 <- nomes_cnae_secao(df1, col1)
df2
}
