% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomes_cnae_subclasse.R
\encoding{UTF-8}
\name{nomes_cnae_subclasse}
\alias{nomes_cnae_subclasse}
\title{Nomes Classificação Nacional de Atividades Econômicas - Subclasse}
\usage{
nomes_cnae_subclasse(tabela, campo)
}
\arguments{
\item{tabela}{Dataframe: a tabela para adicionar os nomes das subclasses da CNAE}

\item{campo}{Caractere: a coluna com os códigos das subclasses da CNAE}
}
\value{
Adiciona uma coluna com as descrições das subclasses da CNAE em uma tabela que possui apenas os códigos.
}
\description{
Procura em dataframe por coluna com códigos CNAE subclasse e adiciona uma coluna com os nomes relacionados.
}
\examples{
df1 <- data.frame(codigos_cnae_subclasse = c("0111301", "4520008", "9900800"))
df2 <- nomes_cnae_subclasse(df1, "codigos_cnae_subclasse")
df2

df3 <- data.frame(codigos_cnae_subclasse = c(111301, 4520008, 9900800))
df4 <- nomes_cnae_subclasse(df3, "codigos_cnae_subclasse")
df4
}
